requirements.txt
-i https://pypi.org/simple
alembic==1.0.9
certifi==2019.3.9
chardet==3.0.4
click==7.0
fakeredis==1.0.3
flask-basicauth==0.2.0
flask-caching==1.7.1
flask-cors==3.0.7
flask-migrate==2.4.0
flask-sqlalchemy==2.3.2
flask==1.0.2
gevent==1.4.0
greenlet==0.4.15 ; platform_python_implementation == 'CPython'
gunicorn==19.9.0
hashids==1.2.0
idna==2.8
itsdangerous==1.1.0
jinja2==2.10.1
mako==1.0.9
markupsafe==1.1.1
psycopg2-binary==2.8.2
python-binary-memcached==0.28.0
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
redis==3.2.1
requests==2.21.0
six==1.12.0
sortedcontainers==2.1.0
sqlalchemy==1.3.3
typing==3.6.6
uhashring==1.1
urllib3==1.24.2
werkzeug==0.15.2
